Fix evsys user[m] register locations

+++ pac/atsame54n/src/evsys/user.rs
@@ -10,7 +10,7 @@ impl R {
     #[doc = "Bits 0:5 - Channel Event Selection"]
     #[inline(always)]
     pub fn channel(&self) -> ChannelR {
-        ChannelR::new((self.bits & 0x3f) as u8)
+        ChannelR::new(self.bits & 0x3f)
     }
 }
 impl W {
@@ -24,18 +24,18 @@ impl W {
 #[doc = "User Multiplexer n\n\nYou can [`read`](crate::Reg::read) this register and get [`user::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`user::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct UserSpec;
 impl crate::RegisterSpec for UserSpec {
-    type Ux = u32;
+    type Ux = u8;
 }
 #[doc = "`read()` method returns [`user::R`](R) reader structure"]
 impl crate::Readable for UserSpec {}
 #[doc = "`write(|w| ..)` method takes [`user::W`](W) writer structure"]
 impl crate::Writable for UserSpec {
     type Safety = crate::Unsafe;
-    const ZERO_TO_MODIFY_FIELDS_BITMAP: u32 = 0;
-    const ONE_TO_MODIFY_FIELDS_BITMAP: u32 = 0;
+    const ZERO_TO_MODIFY_FIELDS_BITMAP: u8 = 0;
+    const ONE_TO_MODIFY_FIELDS_BITMAP: u8 = 0;
 }
 #[doc = "`reset()` method sets USER[%s]
 to value 0"]
 impl crate::Resettable for UserSpec {
-    const RESET_VALUE: u32 = 0;
+    const RESET_VALUE: u8 = 0;
 }
